perf: changed flat list in single user profile
