+++ src/screens/HomeScreen/UserRow.js
@@ -5,18 +5,10 @@ import { Avatar } from 'react-native-elements'
 import MaterialCommunityIcons from 'react-native-vector-icons/MaterialCommunityIcons'
 import DoubleTap from 'react-native-double-tap'
 
-import { getRandomLightColor } from '../../helpers/getRandomLightColor'
+import { SmallPill } from '../../components/SmallPill'
+import { getColorsArray } from '../../helpers/getColorsArray'
 import styles from './styles'
 import { editUserInfo, _addLike, _removeLike } from '../../store/userReducer'
-import { firebase } from '../../firebaseSpecs/config'
-
-export const getColorsArray = (num) => {
-  const colors = new Array(num)
-  for (let i = 0; i < colors.length; i++) {
-    colors[i] = getRandomLightColor()
-  }
-  return colors
-}
 
 export default function UserRow({ item, navigation }) {
   const likes = useSelector((state) => state.user.likes)
@@ -45,6 +37,12 @@ export default function UserRow({ item, navigation }) {
     }
   }
 
+  const renderInterests = (interests) => {
+    return interests.map((item, index) => (
+      <SmallPill key={index} backgroundColor={colors[index]} text={item} />
+    ))
+  }
+
   return (
     <TouchableOpacity
       onPress={() =>
@@ -70,16 +68,9 @@ export default function UserRow({ item, navigation }) {
               )}
             </DoubleTap>
           </View>
-          <FlatList
-            style={styles.interestsContainer}
-            data={item.interests}
-            renderItem={({ item, index }) => (
-              <View style={styles.interest} backgroundColor={colors[index]}>
-                <Text style={styles.interestText}>{item}</Text>
-              </View>
-            )}
-            keyExtractor={(item, index) => (item + index).toString()}
-          />
+          <View style={styles.interestsContainer}>
+            {renderInterests(item.interests)}
+          </View>
         </View>
       </View>
     </TouchableOpacity>
